Refactor error handling. Move Cadence templates to separate file
